% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcure.r
\docType{data}
\name{leukaemia}
\alias{leukaemia}
\title{Data from leukaemia patients with bone marrow transplants}
\format{
A data.frame object with columns:
\describe{
 \item{\code{time}:}{Relapse Free Survival Time.}
 \item{\code{cens}:}{Relapse Indicator with 1-Relapsed, 0-Disease Free.}
 \item{\code{transplant}:}{Bone marrow transplant: 0-Allogeneic, 1-Autologous.}
}
}
\source{
Kersey JH, Weisdorf D, Nesbit ME, LeBien TW, Woods WG, McGlave PB, Kim TRUE, Vallera DA, Goldman AI, Bostrom B, Hurd D, Ramsay NKC. Comparison of autologous and allgeneic bone marrow transplantation for treatment of high-risk refractory acute lymphoblastic leukaemia. New England Journal of Medicine 1987; 317:461– 467.
}
\usage{
data(leukaemia)
}
\description{
This data set is used in the examples of this package.
}
\examples{
data(leukaemia)
}
\keyword{datasets}
