% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcure-package.R
\docType{package}
\name{mixcure-package}
\alias{mixcure-package}
\title{Mixture cure models}
\description{
Fit various mixture cure models
}
\details{
This package fits various mixture cure model using existing R packages.
}
\references{
Peng, Y. and Yu, B. Cure Models: Methods, Applications, and Implementation.
CRC/Chapman & Hall, 2020

Peng, Y. and Taylor, J. M. G. Cure models. In Klein, J.,
van Houwelingen, H., Ibrahim, J. G., and Scheike, T. H., editors, Handbook
of Survival Analysis, Handbooks of Modern Statistical Methods series,
chapter 6, pages 113-134. Chapman & Hall, Boca Raton, FL, USA, 2014

Peng, Y. Fitting semiparametric cure models. Computational Statistics & Data
Analysis, 41: 481-490, 2003
}
\author{
Yingwei Peng
}
