% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcure.r
\name{print.mixcure}
\alias{print.mixcure}
\title{Print method for mixture cure models}
\usage{
\method{print}{mixcure}(x, ...)
}
\arguments{
\item{x}{a mixcure object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for mixture cure models
}
\examples{

data(leukaemia)
z1 = mixcure(Surv(time, cens) ~ transplant, ~ transplant, data = leukaemia,
savedata = TRUE)
print(z1)

}
\seealso{
mixcure
}
\author{
Yingwei Peng
}
